requirements.txt
Flask==2.0.1
opencv-python==4.5.3.56
pdf2image==1.16.0
Pillow==8.3.1
numpy==1.21.2